backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
Flask-JWT-Extended>=4.6.0
Flask-Cors>=4.0.1
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
email-validator>=2.0.0
